# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mistri APIs",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/customer/get-cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get All Cars",
                "operationId": "getCars",
                "consumes": [
                    "application/xml",
                    "application/json"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/customer/add-customer-car": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Add customer's car",
                "operationId": "add_customer_car",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "car_id",
                        "in": "formData",
                        "description": "Car Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "millage",
                        "in": "formData",
                        "description": "Millage",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "vehicle_no",
                        "in": "formData",
                        "description": "Vehicle Number",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "insurance",
                        "in": "formData",
                        "description": "Insurance",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "year",
                        "in": "formData",
                        "description": "Year",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/customer/remove-customer-car": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Delete customer's car",
                "operationId": "remove_customer_car",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "car_id",
                        "in": "formData",
                        "description": "Car Id",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "vehicle_no",
                        "in": "formData",
                        "description": "Vechile No.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/customer/get-customer-car": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get customer's car",
                "operationId": "getCustomerCar",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}